pyproject.toml
[project]
name = "codesentinel"
version = "0.1.0"
description = "Automated code review RL environment built on the OpenEnv framework for the Meta PyTorch Hackathon"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openenv-core",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.7.0",
    "websockets>=12.0",
    "openai>=1.30.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["server"]